Fixed issue with use of corr_terms in likelihood function, fixed issue with confidence envelopes in trackplot, added posterior samples to deltaRA-deltaDEC plot, removed the jd_tt parameter and added the ref_epoch parameter to generate_plots, added fileprefix parameter to system class, simplified the extraction of samples by using samples_equal from dynesty, added Github Actions workflow, updated .gitignore, updated package __init__, added unit tests to tests folder, removed all binary files that are not required by the repo, tested compatibility with Python 3.11
