+++ .gitignore
@@ -1,5 +1,22 @@
+__pycache__
+*.pyc
+.DS_Store
+docs/_build
+docs/.ipynb_checkpoints/
+build/
+dist/
+backtracks.egg-info/
+.coverage*
+coverage.xml
+htmlcov/
+.tox/
 *.ipynb*
 bgstar_emcee.py
-__pycache__
 *beta_pic*
 *SAO*
+.pyproject.toml.sw
+.pytest_cache/
+tests/HD_131399_A_*.png
+tests/HD_131399_A_dynestyrun_results.pkl
+tests/gaia_query_6204835284262018688.fits
+tests/dynesty.save
