+++ backtracks/__init__.py
@@ -1,0 +1,8 @@
+from backtracks.backtracks import system
+
+__author__ = "William Balmer"
+__license__ = "BSD-3"
+__version__ = "0.3.1"
+__maintainer__ = "William Balmer"
+__email__ = "wbalmer1@jhu.com"
+__status__ = "Development"
