+++ pyproject.toml
@@ -13,7 +13,7 @@ authors = [
 ]
 description = "Python package to fit relative astrometry with background star motion tracks."
 readme = "README.md"
-requires-python = ">=3.9,<3.10"
+requires-python = ">=3.9,<3.12"
 dynamic = ["dependencies"]
 classifiers = [
     "Programming Language :: Python :: 3",
@@ -21,9 +21,6 @@ classifiers = [
 
 [tool.setuptools.dynamic]
 dependencies = { file = ["requirements.txt"] }
-
-[tool.setuptools.packages.find]
-where = ["src"]
 
 [project.urls]
 Homepage = "https://github.com/wbalmer/backtrack"
