--- src/backtracks.egg-info/SOURCES.txt
@@ -1,32 +1,0 @@
-.gitignore
-.pyproject.toml.swp
-LICENSE
-README.md
-dynesty.save
-pyproject.toml
-requirements.txt
-dist/backtracks-0.3.0-py3-none-any.whl
-dist/backtracks-0.3.0.tar.gz
-src/backtracks/__init__.py
-src/backtracks/backtrack.py
-src/backtracks/bailer-jones_edr3.csv
-src/backtracks/plotting.py
-src/backtracks/utils.py
-src/backtracks.egg-info/PKG-INFO
-src/backtracks.egg-info/SOURCES.txt
-src/backtracks.egg-info/dependency_links.txt
-src/backtracks.egg-info/requires.txt
-src/backtracks.egg-info/top_level.txt
-tests/HD_131399_A_bjprior_backtrack.pdf
-tests/HD_131399_A_corner_backtrack.pdf
-tests/HD_131399_A_dynestyrun_results.pkl
-tests/HD_131399_A_evidence_backtrack.pdf
-tests/HD_131399_A_model_backtrack_radec.pdf
-tests/HD_131399_A_model_backtrack_seppa.pdf
-tests/HD_131399_A_nearby_gaia_distribution.pdf
-tests/bailer-jones_edr3_prior_summary.csv
-tests/dynesty.save
-tests/gaia_query_6204835284262018688.fits
-tests/hd131339a.py
-tests/scorpions1b_orbitizelike.csv
-tests/test_plotting.py
