--- src/backtracks.egg-info/requires.txt
@@ -1,13 +1,0 @@
-astropy
-astroquery
-corner
-dynesty
-matplotlib
-numpy
-novas
-novas_de405
-orbitize
-pandas
-seaborn
-schwimmbad
-scipy
